+++ extensions/dotnav/src/solutionSearch/index.ts
@@ -4,5 +4,5 @@ export * from './searchEngine';
 export * from './searchDiskStore';
 export * from './searchCommands';
 export * from './searchStatusBar';
-
-
+export * from './searchUpdateCoordinator';
+export * from './latestQueryScheduler';